refactor: reorganize Jotai atom structure and enhance swap analytics management
- Introduced a mandatory structure for organizing Jotai atoms, distinguishing between global and feature-specific state management.
- Added new swap analytics state management in the marketV2 context, consolidating analytics settings for amount entry type and slippage.
- Removed the outdated swapAnalytics atom from the SwapPanel component, streamlining the analytics logic.
- Updated the useSwapAnalytics hook to utilize the new atom structure, improving clarity and maintainability of analytics tracking.
