+++ CLAUDE.md
@@ -314,10 +314,51 @@ OneKey is an open-source multi-chain crypto wallet with a monorepo architecture 
 - Minimize platform-specific code by keeping common logic separate
 
 ### State Management
-- Use Jotai atoms for state
-- Create context providers for complex state
-- Implement selectors for derived state
-- Keep platform-specific state isolated
+
+**Jotai Atom Organization - MANDATORY STRUCTURE**:
+
+**Global State Atoms** (for app-wide, persistent state):
+- **Location**: `packages/kit-bg/src/states/jotai/atoms/`
+- **Usage**: Global settings, account state, hardware state, currency, etc.
+- **Pattern**: Use `globalAtom` and `EAtomNames` for standardization
+- **Examples**: `settings.ts`, `account.ts`, `hardware.ts`, `currency.ts`
+
+**Feature-Specific State Atoms** (for localized functionality):
+- **Location**: `packages/kit/src/states/jotai/contexts/[功能名]/atoms.ts`
+- **Usage**: Feature-specific state that may be shared across components within that feature
+- **Pattern**: Use `contextAtom` from `createJotaiContext` for consistency
+- **Structure**:
+  ```
+  contexts/
+  ├── marketV2/
+  │   ├── atoms.ts     - State definitions
+  │   ├── actions.ts   - State operations
+  │   └── index.ts     - Exports
+  ├── swap/
+  │   ├── atoms.ts
+  │   ├── actions.ts
+  │   └── index.ts
+  ```
+
+**FORBIDDEN Atom Patterns**:
+- ❌ **NEVER** create atom directories under `packages/kit/src/views/`
+- ❌ **NEVER** create standalone atom files in component directories
+- ❌ **NEVER** mix `globalAtom` and `contextAtom` patterns without architectural justification
+
+**Atom Selection Guidelines**:
+1. **Use globalAtom when**:
+   - State needs persistence across app restarts
+   - State is used across multiple major features
+   - State affects the entire application (settings, authentication, etc.)
+   - Located in `packages/kit-bg/src/states/jotai/atoms/`
+
+2. **Use contextAtom when**:
+   - State is specific to a feature or module
+   - State is temporary/session-based
+   - State is shared within related components of a feature
+   - Located in `packages/kit/src/states/jotai/contexts/[name]/atoms.ts`
+
+**IMPORTANT**: These are the ONLY two atom patterns used in the project. Do not create custom atom patterns or use plain Jotai atoms outside of these established structures.
 
 ## Coding Patterns and Best Practices
