+++ packages/kit/src/states/jotai/contexts/marketV2/atoms.ts
@@ -5,6 +5,10 @@ import {
   atom,
   createJotaiContext,
 } from '@onekeyhq/kit/src/states/jotai/utils/createJotaiContext';
+import {
+  EAmountEnterType,
+  ESlippageSetting,
+} from '@onekeyhq/shared/src/logger/scopes/dex/types';
 import type { IMarketWatchListDataV2 } from '@onekeyhq/shared/types/market';
 import type { IMarketTokenDetail } from '@onekeyhq/shared/types/marketV2';
 
@@ -48,6 +52,20 @@ export const { atom: selectedNetworkIdAtom, use: useSelectedNetworkIdAtom } =
 
 export const { atom: selectedMarketTabAtom, use: useSelectedMarketTabAtom } =
   contextAtom<string>('trending');
+
+// SwapPanel Analytics Atoms
+export interface ISwapAnalyticsState {
+  amountEnterType: EAmountEnterType;
+  slippageSetting: ESlippageSetting;
+}
+
+const initialSwapAnalyticsState: ISwapAnalyticsState = {
+  amountEnterType: EAmountEnterType.Manual,
+  slippageSetting: ESlippageSetting.Auto,
+};
+
+export const { atom: swapAnalyticsAtom, use: useSwapAnalyticsAtom } =
+  contextAtom<ISwapAnalyticsState>(initialSwapAnalyticsState);
 
 const INIT = Symbol('INIT');
 export const marketWatchListV2Atom = memoizee(() =>
